% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEswan.R
\name{DEswan}
\alias{DEswan}
\title{Differential Expression - Sliding Window ANalysis (DE-SWAN)}
\usage{
DEswan(data.df, qt, window.center, buckets.size, covariates)
}
\arguments{
\item{data.df:}{a data Frame - columns are variables and rows are samples (same order as qt and covariates)}

\item{qt:}{a Vector - quantitative trait for DE-SWAN (same order as data.df and covariates)}

\item{window.center:}{an optional Vector - window centers. Default is quantile(qt,  probs = seq(.1,.9,.1))}

\item{buckets.size:}{an optional Numeric - size of each bucket. Default is set to (max(range(qt))-min(range(qt)))/2}

\item{covariates:}{a data Frame. Columns are variables and rows are samples (same order as qt and covariates)}
}
\value{
list of 2 data Frames: $p  =  pvalues of qT and covariates for each variable and each window.center and $coefficients  =  coefficients of qT and covariates for each variable and each window.center
}
\description{
Differential Expression - Sliding Window ANalysis (DE-SWAN)
}
\examples{
res.DEswan=DEswan(data.df = agingplasmaproteome[,-c(1:3)],
               qt = agingplasmaproteome[,1],
               window.center = seq(40,100,10),
               buckets.size = 10,
               covariates = agingplasmaproteome[,c(2:3)])
head(res.DEswan$p)
head(res.DEswan$coeff)

}
