% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.DEswan.R
\name{reshape.DEswan}
\alias{reshape.DEswan}
\title{Reshape DE-SWAN output}
\usage{
reshape.DEswan(res.DEswan, parameter, factor)
}
\arguments{
\item{res.DEswan:}{a data Frame - output from DEswan}

\item{parameter:}{an optional numeric - 1 for pvalue or 2 for coefficients. Default is 1}

\item{factor:}{an optional character string - qt or covariates in DEswan. Default is qt.}
}
\description{
Reshape DE-SWAN output
}
\examples{
# list of factor
unique(res.DEswan$p$factor)
unique(res.DEswan$coeff$factor)
res.DEswan.wide.p=reshape.DEswan(res.DEswan)
res.DEswan.wide.p=reshape.DEswan(res.DEswan,parameter = 1,factor = "qt")
res.DEswan.wide.coeff=reshape.DEswan(res.DEswan,parameter = 2,factor = "qt")



}
