% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsignif.DEswan.R
\name{nsignif.DEswan}
\alias{nsignif.DEswan}
\title{Calculate number of significant variables for a set of pvalues thresholds}
\usage{
nsignif.DEswan(res.DEswan.wide, thresholds)
}
\arguments{
\item{res.DEswan.wide:}{a data Frame - pvalues or qvalues from DEswan in a wide format}

\item{thresholds:}{a Vector - p or q-values thresholds. Default is c(0.05,1e-2,1e-3,1e-4,1e-5)}
}
\description{
Calculate number of significant variables for a set of pvalues thresholds
}
\examples{
res.DEswan.wide.p.signif=nsignif.DEswan(res.DEswan.wide.p)



}
