% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q.DEswan.R
\name{q.DEswan}
\alias{q.DEswan}
\title{caculate qvalues for DE-SWAN output}
\usage{
q.DEswan(res.DEswan.wide, method)
}
\arguments{
\item{res.DEswan.wide.p:}{a data Frame - pvalues from DEswan in a wide format}

\item{method:}{method for pvalues adjustment. Default is BH}
}
\description{
caculate qvalues for DE-SWAN output
}
\examples{
res.DEswan.wide.q=q.DEswan(res.DEswan.wide.p,method="BH")
res.DEswan.wide.q=q.DEswan(res.DEswan.wide.p,method="Bonferonni")



}
