% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.DEswan.R
\name{reshape.DEswan}
\alias{reshape.DEswan}
\title{Reshape DE-SWAN output}
\usage{
reshape.DEswan(res.DEswan, parameter, factor)
}
\arguments{
\item{res.DEswan:}{a data Frame - output from DEswan}

\item{parameter:}{an optional numeric - 1 for pvalue or 2 for coefficients. Default is 1}

\item{factor:}{an optional character string - qt or covariates in DEswan. Default is qt.}
}
\description{
Reshape DE-SWAN output
}
\examples{
res.DEswan=DEswan(data.df = agingplasmaproteome[,-c(1:3)],
qt = agingplasmaproteome[,1],
window.center = seq(40,100,10),
buckets.size = 10,
covariates = agingplasmaproteome[,c(2:3)])
head(res.DEswan$p)
head(res.DEswan$coeff)
res.DEswan.wide.p=reshape.DEswan(res.DEswan)
res.DEswan.wide.p=reshape.DEswan(res.DEswan,parameter = 1,factor = "qt")
head(res.DEswan.wide.p)
res.DEswan.wide.coeff=reshape.DEswan(res.DEswan,parameter = 2,factor = "qt")
head(res.DEswan.wide.coeff)
}
